requirements.txt
langchain
langchain-openai
langchain-community
langgraph
textual
pandas
python-dotenv
kaggle
langchain-experimental
duckduckgo-search
matplotlib
seaborn
scikit-learn
xgboost
lightgbm
optuna
shap
jinja2
langgraph.checkpoint.sqlite
oumi